Lets not shit where we eat and use a linter
