+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "./node_modules/.bin/mocha ./test/**/*[sS]pec.js",
-    "karma": "./node_modules/.bin/karma start"
+    "karma": "./node_modules/.bin/karma start",
+    "lint": "./node_modules/.bin/eslint **.js"
   },
   "repository": {
     "type": "git",
@@ -28,6 +29,10 @@
   "devDependencies": {
     "babel-preset-es2015": "^6.14.0",
     "chai": "^3.5.0",
+    "eslint": "^3.4.0",
+    "eslint-config-standard": "^6.0.0",
+    "eslint-plugin-promise": "^2.0.1",
+    "eslint-plugin-standard": "^2.0.0",
     "karma": "^1.2.0",
     "karma-chai": "^0.1.0",
     "karma-chrome-launcher": "^2.0.0",
