+++ server.js
@@ -1,11 +1,1 @@
-var koa = require('koa');
-var app = koa();
-
-console.log("Koa listening on localhost:3000")
-
-app.use(function *(){
-  this.body = 'Hello World';
-});
-
-app.listen(3000);
-
+module.exports = require('./src/app/app.js');
